vmtests: skip Focal deploying Centos70 ScsiBasic
In Focal, there is a bug (#1859858) in udev which returns a
truncated value for ID_SERIAL on some scsi devices; this does
not break dname rules when the ephemeral environment matches
the target system, however if Focal deploys Bionic or in this
case Centos70, the udev in the target system produces a
different ID_SERIAL value which prevents the dname rules from
matching.  Skip this test until we get a udev fix.