clear-holders: ensure we wipe device even if multipath enabled not not mp
On Ubuntu releases where multipath is enabled but not in-use some devices
would fail to be wiped due to logic bug.  This was found in the
BcacheBasic test.  Refactor this logic to ensure that we wipe either
the in-use multipath dm device or the block device value originally
supplied to the wipe function.