vmtests: bump fixby date out and fix false positive when date passes
The fix for vlan is still pending, however we've observed a false
positive where vmtest would indicate that the test was passing when
it actually was not. This is because skip_by_date will run the
decorated function *after* the fixby date has passed. The
implementation was 'pass' which does pass but did not actually run
the test case.

Fix this by calling to the super class.

LP: #1855148