clear-holders: umap the parent mpath to wipe the underlying partitions
In some multipath scenarios access to the underlying partitions is
not possible without removing the parent multipath mapping.  This can
only be done if all of the mpath partitions have been unmapped.
Detect when this is the case and remove the parent mpath mapping
as well.  This fixes vmtest failures found with test_multipath.py