install-grub: refactor uefi partition/disk searching
Refactor the code used to determine the partition and disk where the
ESP resides.  There were a few issues:

 - Disks and partitions are of type 'block' which requires the use of
   [ -b ] or [ -e ]; [ -f ] returns false for block devices.
 - The use of lsblk -on pkname exposed a bug in util-linux where pkname
   did not return a value.

This branch adds additional debug logging and safety checks for any
calculated value and fails on empty values.  We replace use of lsblk
with use of sysfs to calculate the parent disk of the partition.  Add
a vmtest to verify the path where we use efibootmgr to create an entry.

LP: #1862846