vmtest: Fix test_basic.py to run on s390x
The TestBasic currently fails on s390x due to usage of virtio-disks
and an "empty" virtio-scsi device controller. There is an existing
bug in s390x firmware that does not like an "empty" virtio-scsi
controller. Until that bug is fixed, skip TestBasic on s390x and
instead make TestScsiBasic work. This requires capturing output of
/dev/disk/by-path as s390x fstab entries prefer by-path as paths on
s390x are stable[1].

1. https://wiki.ubuntu.com/FSTAB

LP: #1866663