storage_config: Add 'extra_options' parameter to allow custom mkfs
Allow users to provide additional flags to the mkfs command used to
create a filesystem during storage configuration.  The extra_options
value is a string that will be appended to the constructed mkfs
command.  The value will be split on whitespace to show up as
individual arguments to the mkfs command.

LP: #1869069