storage: enable and use multipath during storage configuration
Curtin has attempted to avoid using multipath enabled during installation
as that requires reworking a number of areas with in curtin that use Linux
tooling which is not multipath friendly.  In particular, partitions on
multipath devices appear as disk not as partitions and this makes programs
confused.  However, some scenarios just cannot be handled sanely without
multipath devices enabled.  One such scenario is creating an LVM root on
top of multipath devices.  This branch adds a vmtest scenario to recreate
this deployment configuration.  To resolve the issue curtin will now keep
multipath enabled if available and adjust partition handling to make use
of 'kpartx' and 'dmsetup' to create, manipulate and calculate partition
related information.  There was quite a bit of fallout with this change
so this branch includes a large amount of changes to handle this all the
way back to Xenial.  Xenial and Bionic do not enable multipath by default,
and curtin does not enable multipath in the ephemeral environment, rather
if multipath is presnet in the ephemeral environment then curtin will
make use of it.

Additional changes needed/used to resolve this issue:

- Add --shutdown-plan to curtin clear-holders CLI tool for dumping the
  clear-holders plan of action for a given storage configuration.
- Add --probe-data to curtin block-discover CLI to dump the probert probe
  data out to a file and skip converting to a storage config.
- clear-holders will reload multipath maps to ensure mp is active.
- Add use of --filter to lvm commands to only use mpath devices if mp is
  enabled.
- curthooks now modifies target /etc/lvm.conf with a devices filter if mp
  is enabled.
- Fix dname for mpath devices, using dm-uuid and handling the oddity that
  multipath partitions prefix their dm-uuid with part-X-.
- If multipath enabled, don't hard code root devices to mpath0

LP: #1869075