Fix handing of reusing msdos partitions and flags
Disks with an MSDOS partition table and one partition with the
'bootable' flag set exposed a parsing error in block.sfdisk_info;
further digging revealed that curtin was not setting the bootable
flag on MSDOS partitions when required.  Not setting isn't fatal,
grub still boots the partition, but we should set the flag if told
to do so.  Additionally storage-config conversion of probe data
where the MSDOS boot flag is set was missing.  The following changes
are done to fix this issue

- switch to sfdisk --json output and use/introduce util.load_json()
- block-meta: update parted command to set boot flag if present
- block-discover: prefer ID_PART_ENTRY_FLAGS when present, and 0x80
- Add vmtest to verify reuse of msdos partition and bootable flag
- Fix boot, extended, logical flag verfication
- Fix extended partition size verification

LP: #1875903