vmtests: fix PreservePartitionWipeVg storage config
This failure

  An error occured handling 'disk-sda': ValueError - disk 'disk-sda'
  does not have correct partition table or cannot be read, but preserve
  is set to true. cannot continue installation.

happens because the storage config for the main disk has both wipe and
preserve settings enabled resulting in the primary disk being wiped.

Curtin can either wipe a disk or preserve it but we cannot do both (unlike
partitions where we can keep the partition location, flag, size, etc,
but wipe the data).  This branch

 - Fixes the storage config in the vmtest
 - Adds a RuntimeException indicating this config is invalid
 - Adds unittests for get_device_paths_from_storage_config