tools/jenkins-runner: shuffle test-cases to randomize load
Enabling jenkins-runner to enumerate all of the tests to be
run and then run them through 'shuf' which will randomize
each test class for selection by the parallel threads of
execution. This will reduce the likelihood that we have
four executions of a Bcache or Raid or other heavy storage
which is what we get with current module level parallelism.

This behavior can be disabled by exporting:

CURTIN_VMTEST_SHUFFLE_TESTS=0

It is enabled by default.