curtainer: mask the zfs-mount and zfs-share services
curtin depends on zfsutils-linux via probert-storage, but zfsutils-linux
can't be installed in an unprivileged container as it fails to start the
zfs-mount and zfs-share services as /dev/zfs is missing. We do not
actually need ZFS to work in the container, so the problem can be worked
around by masking the services before the package is installed.