fix verification of vtoc partitions
Currently curtin verifies the size and flags of a partition by
using sfdisk to inspect it and checking the results. But that
doesn't work on a dasd's vtoc partition table, so this branch
renames parition_verify to parition_verify_sfdisk, adds
parition_verify_fdasd and calls the latter when ptable ==
'vtoc'. The branch does a little more than is strictly necessary,
perhaps, but the other stuff will come in useful soon, I promise
:)

LP: #1899471