storage_config: handle some FBA dasd oddities
The storage_config handling of dasds only accounted for ECKD
dasds. FBA dasds need different handling -- in many ways they are
more like a regular disk (that can only be formatted with a DOS
partition table, but this code doesn't need to handle that), but
the kernel creates a magic fake partition for a FBA dasd with no
DOS partition table and we want to pretend that doesn't exist.