install_grub: Fix bootloader-id for RHEL systems, must be redhat
In 7310b4fe61465, the port from shell dropped a change where the boot-id
value for Redhat family OSes is the value 'redhat' not the os-release ID
value (which is rhel, on RHEL systems).

Add this change back to install_grub and fix unittests that fail with
'rhel' instead of 'redhat' in os-release values.  Hardcode 'redhat'
paths in expected values.

LP: #1906543