vmtest: fix image-sync after maas URL stream rename
Image syncing for vmtest has been failing since MAAS renamed their
daily image stream to stable.  In an effort to keep the 'daily' URL
valid the website redirects this to 'stable' path.  However the content
was also modified and includes references to files and URLs which have
the 'stable' name included.  This broke vmtest image sync which expected
the 'daily' string to be present in filenames and content.

The observable result was that curtin would attempt to find content it
had just *synced* to /srv/images but when searching, it would load up
the cached content files which only referenced the 'daily' stream data
and would indicate that the target file was missing.

This branch moves vmtest to using the 'stable' maas image stream
directly.  This value is controllable via an environment variable,
MAAS_IMAGE_STREAM, if set will be used to compose the IMAGE_SRC_URL
and the STREAM_BASE variable which is used to find cached content in
/srv/images/.vmtest_data/ directory.

LP: #1908543