Fix NVMe validation for namespaces with UUID
NVMe namespaces are only verified if the wwn comes from either the
NGUID or EUI64. With NVMe 1.3 a third ID, UUID is now[1] supported for a
namespace. If it exists it takes precedence and the wwid in sysfs will
populate with that uuid and use a uuid.XXXXX format, which does not
validate correctly with the filter in curtin.

/sys/devices/virtual/nvme-subsystem/nvme-subsys0/\
nvme0/nvme0n1/wwid:uuid.289acd66-794d-439d-abd0-0396f6382936

1. https://lore.kernel.org/patchwork/patch/808637/

LP: #1925399