Extend the 'format' schema for unknown fstypes
Permit arbitrary fstype values if preserve == true
Goal here is to enumerate more filesystem formats, ones that blkid
recognizes, but not claim in any way that we can create them.
This is useful for presenting existing filesystem partitions to users
in an installer, as otherwise they just get unhelpfully marked as
'unknown'.