tests: update to demonstrate python-apt functional
Prior to pulling in python{,3}-apt functionality, show a simple set of
tests that demonstrate that it is functional in the test environments.

The noproxy script has been removed from tox.ini because we need to
set sitepackages=True to get python-apt, but doing so means we need to
find nose with 'python -m'.  Also, it's safe to remove 'noproxy'
becuase the operation it was doing was a no-op for tox anyhow (those
proxy variables are not getting passed thru unless we set them
manually in tox.ini)

With noproxy unused by tox, it had no users, so remove it entirely.