block_meta: call realpath on partition node returned by sfdisk
sfdisk --json on a device mapper block device will "helpfully" return
paths like /device/mapper/mpatha-part1 for the "node" of each partition.
This doesn't work so well when you then assume that the basename of the
returned path is the kernel name for the device.

This fixes reusing a partition on a multipath disk.