block_meta_v2: change how we invoke sfdisk again, restore partprobe call
It's a bit disappointingly empirical but some multipath vmtests fail
with the current sfdisk arguments with "device busy" type errors.  I
think this points to a bug somewhere else but the current partition code
works in this situation so massage things so that the new code does too.

Now we pass --no-tell-kernel to sfdisk, we do need to tell the kernel
about the changes so put back a call to partprobe to do this. This is
actually needed independent of the above change, because it seems sfdisk
with default arguments does not cause the kernel to remove the device
node for a deleted partition!