examples: stop assuming curtin accounts for overhead of logical partitions
Curtin currently has the slightly surprising (to me at least) behaviour
of enlarging an extended partition to account for the overhead that
comes from the fact that logical partitions have to have gaps between
them. I do not intend to preserve this in v2 partitioning (and will
handle this in subiquity) so change the two example files that depend on
this.

Also correct an offset in examples/tests/multipath-reuse.yaml.