examples: even more tweaks for v2
bcache-partitions.yaml: remove the ptable setting on a drive that will
be part of a bcache. A quirk of v1 partitioning is that setting ptable:
gpt on a disk but then creating no partitions will leave the disk with
no partition table.

lvm_iscsi.yaml: enlarge some partitions, one for / and one because of
the logical partition overhead.

mirrorboot.yaml: make / slightly larger.