usage: fix help that showed invalid JSON in examples
The usage showed examples to override the stages that are
executed when running curtin install. That said, JSON only
accepts strings surrounded by double quotes and the examples use
single quotes. This leads to invalid JSON syntax if the examples
are copied.

Fixed by inverting the quotes used in the examples so that the
examples are valid.