tests/apt: use tmpdir as target
This refactor helps make way for discontinuing usage of apt-key.
When trying to do it all at once, there were problems with the file path
that needed to be used for the key file.

The major change is that these tests are run with a tmpdir as the
target, instead of '/'.  This is also independently a nice improvement,
since misbehaved tests or code would really try to write to the host
system.

The tests were written assuming apt key files would have full paths, but
the examples (and presumably most real-world usage) were based on
filenames, so move the tests to use file names.

Add back a full path test, since that is supported behavior today.

Also removes the unused _myjoin function.