block_meta_v2: set first-lba if needed when creating a new GPT
sfdisk sets first-lba to 2048 by default, which fails if one of
the partitions has an start smaller than that.  Having a
partition start so early in the disk is a bit strange but it
works if you are preserving a partition (because we preserve
first-lba) so it should probably work when creating one too.