system-install: use --assume-downloaded instead of --no-download
Running $ apt-get install --no-download
fails if the packages were previously "downloaded" using the cdrom
rather than an online repository.

This means we can't reliably pass the option on the second apt-get
install invocation.
Passing the --no-download option to $ curtin system-install  while not
passing it to apt-get feels wrong. Therefore, I've renamed the
system-install option from --no-download to --assume-downloaded.

For package managers that support the --no-download option properly,
they can use it on the second invocation. For apt, we will not pass the
option.

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>