apt-config: stop returning cfg object in translate_old_apt_features
The translate_old_apt_features function mutates its parameter, but also
used to return the translated object. This goes against the rules of
idiomatic Python. Returning a new configuration object without mutating
the original one would require leaning on copy.deepcopy, which is
expensive and not always needed.

Let's stop returning the resulting configuration so the interface is
clear.

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>