udev: fix crash when PARTNAME is invalid utf-8
(We already merged a similar fix in probert)

GPT partition names are basically free-text fields but the udev PARTNAME
field ends up butchered by the kernel when it contains characters
outside the ASCII range.

When using pyudev, the values are all expected to be decodable using the
system's default encoding.

In curtin, it can result in a crash when iterating over all the
properties of a block device, that are mapped by pyudev.

To work around the issue, we now stop iterating over all the values. We
iterate over the keys and only access the value if we need it.

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>
(cherry picked from commit 3969bca45f51057f532272c5666673579e03a59e)