tests: make random_string a staticmethod instead of classmethod
CiTestCase.random_string was decorated with @classmethod.
I guess the purpose was to make the function usable without a CiTestCase
instance. However, @classmethod is used when a function should return an
instance of the class.

This is not what we want. @staticmethod is the right decorator in this
context.

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>
(cherry picked from commit 0131c2dd9f18817a457a8114f27dafee0e651af9)