compat: create module
As curtin is highly dependent on underlying tools, and curtin must
support different runtime environments, we should dynamically handle and
use those tools appropriately.

The immediate problem is availability, or not, of command line flags on
two util-linux utilities.  Add version scanning so that the relevant
code paths can desice what to do.