apt-config: introduce want_deb822() helper
Add a helper function that indicates whether or not we want deb822
sources on the target. This is determined by checking if we are running
Ubuntu 23.10 or later, because this is the release where deb822 sources
are used by default.

(cherry picked from commit f519abebf35a5e52e380e7d67054e9dd68707e6c)