Add support for partitions directly on dm-crypt devices (no LVM)
While not common, this setup is possible. This configuration may be desireable
for anyone that wishes to use dm-crypt without LVM.

There was a proposal some time ago to make this configuration
default on Ubuntu. While that proposal wasn't adopted, I think that this
configuration is something that we should at least support for Desktop / Server
installs for those that manually configure their own devices.