curthooks: nvmeotcp: bring up the network in the initramfs
If the network is required in the initramfs for NVMe over TCP, we now
extract the netplan configuration, generate a set of `ip` commands (plus
dhcpcd if needed) and execute those in the initramfs. This is a bit of a
hack because the netplan configuration supplied by subiquity is provided
using write_files directives. Also, only trivial network configurations
are supported at the moment.

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>