curthooks: nvmeotcp: check if network in initramfs is needed
To decide if we should bring up the network in the initramfs and then
run `nvme connect-all` commands, we now look for mounpoints that have
the `_netdev` option. Those are the ones that require the network to be
up before mounting. If any of those mounpoints corresponds to something
essential for booting (e.g., /, /usr, /var or alike), then we decide
that the network is required in the initramfs.

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>